+++ modules/monitoring_function/variables.tf
@@ -111,3 +111,9 @@ variable "application_insights_custom_name" {
   type        = string
   default     = null
 }
+
+variable "storage_account_enable_advanced_threat_protection" {
+  description = "FAME advanded thread protection (aka ATP) on Function App's storage account"
+  type        = bool
+  default     = false
+}